send_thinkfan_sparkplug/requirements.txt
# MQTT client
paho-mqtt>=1.6.1
# System monitoring
psutil>=5.9.0
# Sparkplug B / Protocol Buffers
protobuf>=3.20.0
# For å laste inn miljøvariabler fra .env-fil
python-dotenv>=1.0.0